Added agent_as_tools sample. Organized the temporal directory structure.
