+++ examples/temporal/workflows/research_bot_workflow.py
@@ -1,6 +1,6 @@
 from temporalio import workflow
 
-from examples.temporal.workflows.research_manager import ResearchManager
+from examples.temporal.workflows.research_agents.research_manager import ResearchManager
 
 @workflow.defn(sandboxed=False)
 class ResearchWorkflow:
